Update checkout, Handle response order
- Add discount price
- Update info check for the user has account

+++ src/main/java/com/techshop/order/dto/voucher/CreateVoucherDto.java
@@ -32,5 +32,5 @@ public class CreateVoucherDto {
 
     @NotNull(message = "Capped at must not be null")
     @Positive(message = "Capped at must greater than 0")
-    private Integer cappedAt;
+    private Long cappedAt;
 }
